Rename and Separate Use Case
